refactor : id, createdAt, updatedAt 값들 base entity로 상속
