+++ src/main/java/ohho/backend/spring/domain/member/entities/Member.java
@@ -1,43 +1,48 @@
-package ohho.backend.spring.domain.exercise.entity;
+package ohho.backend.spring.domain.member.entities;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
-import javax.persistence.Convert;
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 import lombok.Getter;
-import ohho.backend.spring.domain.exercise.enums.ExerciseType;
-import ohho.backend.spring.domain.exercise.enums.Part;
-import ohho.backend.spring.domain.exercise.enums.converter.ExerciseTypeConverter;
-import ohho.backend.spring.domain.exercise.enums.converter.PartConverter;
-import ohho.backend.spring.domain.exerciseHistory.entity.ExerciseHistory;
+import ohho.backend.spring.common.entities.BaseEntity;
+import ohho.backend.spring.domain.exerciseHistory.entities.ExerciseHistory;
+import ohho.backend.spring.domain.group.entities.Group;
+import ohho.backend.spring.domain.partner.entities.Partner;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 
 @Entity
 @Getter
-public class Exercise {
+public class Member extends BaseEntity {
 
     @Id
     @GeneratedValue
     private Long id;
 
-    private String name;
+    private String nickName;
 
-    @Convert(converter = ExerciseTypeConverter.class)
-    private ExerciseType exerciseType;
-
-    @Convert(converter = PartConverter.class)
-    private Part part;
+    private String interestList;
 
     /* 연관관계 */
 
-    @OneToMany(mappedBy = "exercise", cascade = CascadeType.ALL)
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "partner_id")
+    private Partner partner;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "group_id")
+    private Group group;
+
+    @OneToMany(mappedBy = "member", cascade = CascadeType.ALL)
     private List<ExerciseHistory> exerciseHistories = new ArrayList<>();
 
     /* date 컬럼 */
