+++ src/main/java/ohho/backend/spring/domain/partner/entities/Partner.java
@@ -1,5 +1,6 @@
-package ohho.backend.spring.domain.group.entity;
+package ohho.backend.spring.domain.partner.entities;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -9,22 +10,30 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.OneToMany;
+import lombok.AccessLevel;
 import lombok.Getter;
-import ohho.backend.spring.domain.member.entity.Member;
+import lombok.NoArgsConstructor;
+import ohho.backend.spring.common.entities.BaseEntity;
+import ohho.backend.spring.domain.member.entities.Member;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 
 @Entity
 @Getter
-public class Group {
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Partner extends BaseEntity {
 
     @Id
     @GeneratedValue
     private Long id;
 
+    private LocalDate dDay;
+
+    private String goal;
+
     /* 연관관계 */
 
-    @OneToMany(mappedBy = "group", cascade = CascadeType.ALL)
+    @OneToMany(mappedBy = "partner", cascade = CascadeType.ALL)
     private List<Member> members = new ArrayList<>();
 
     /* date 컬럼 */
